Remove projector rendered event
